[STYLE] for test info page
