+++ src/components/TestInfo.jsx
@@ -1,15 +1,16 @@
 import PropTypes from 'prop-types';
 import { useSession } from '../context/SessionContext';
 import useRequest from '../hooks/useRequest';
+import AddIcon from '../assets/icons/add.svg';
+import DownloadIcon from '../assets/icons/download.svg';
 
-function TestInfo({ examName, option }) {
+function TestInfo({ examName }) {
   const { session } = useSession();
   const { response, loading, error } = useRequest(
     session.token,
-    `${process.env.NEXT_PUBLIC_API_URL}/exams/?name=${examName}`,
+    `${process.env.NEXT_PUBLIC_API_URL}/exams?name=${examName}`,
     1,
   );
-
   if (loading) {
     return <div className="loader" />;
   }
@@ -18,39 +19,67 @@ function TestInfo({ examName, option }) {
     return <h3>{error.message}</h3>;
   }
 
+  // PENDING UNTIL THE PART OF DOWNLOAD IS DONE
+  async function downloadTest() {}
+  // PENDING UNTIL THE ASSIGN PAGE IS READY
+  function assignResult() {}
+
   return (
     <section className="test-detail__data">
-      {option}
-      <h1 className="test-detail__data__title">Test Information</h1>
-      <div className="test-detail__data__container">
-        <div className="test-detail__data__element">
-          <strong className="test-detail__data__element__head">Test name</strong>
-          <span className="test-detail__data__element__body">{response.data.name}</span>
-        </div>
-        <div className="test-detail__data__element">
-          <strong className="test-detail__data__element__head">Test short name</strong>
-          <span className="test-detail__data__element__body">{response.data.shortName}</span>
-        </div>
+      <div className="test-detail__data__container--principal">
+        <h1 className="test-detail__data__title">Test Information</h1>
+        {session.user.typeOfUser === 'Bacteriologist' ? (
+          <AddIcon
+            className="test-detail__data__option"
+            onClick={assignResult}
+          />
+        ) : (
+          <DownloadIcon
+            className="test-detail__data__option"
+            onClick={downloadTest}
+          />
+        )}
       </div>
-      <div className="test-detail__data__element">
-        <strong className="test-detail__data__element__head">Description</strong>
-        <span className="test-detail__data__element__body">{response.data.description}</span>
-      </div>
-      <div className="test-detail__data__element">
-        <strong className="test-detail__data__element__head">Indications</strong>
-        <span className="test-detail__data__element__body">{response.data.indications}</span>
-      </div>
-      <div className="test-detail__data__element">
-        <strong className="test-detail__data__element__head">Schedule minimum days</strong>
-        <span className="test-detail__data__element__body">{`${response.data.scheduleDays} days`}</span>
-      </div>
+      {response && response.data ? (
+        <>
+          {response.data.map((test) => (
+            <>
+              <div className="test-detail__data__container">
+                <div className="test-detail__data__element">
+                  <strong className="test-detail__data__element__head">Test name</strong>
+                  <span className="test-detail__data__element__body">{test.name}</span>
+                </div>
+                <div className="test-detail__data__element">
+                  <strong className="test-detail__data__element__head">Test short name</strong>
+                  <span className="test-detail__data__element__body">{test.shortName}</span>
+                </div>
+              </div>
+              <div className="test-detail__data__element">
+                <strong className="test-detail__data__element__head">Description</strong>
+                <span className="test-detail__data__element__body">{test.description}</span>
+              </div>
+              <div className="test-detail__data__container">
+                <div className="test-detail__data__element">
+                  <strong className="test-detail__data__element__head">Indications</strong>
+                  <span className="test-detail__data__element__body">{test.indications}</span>
+                </div>
+                <div className="test-detail__data__element">
+                  <strong className="test-detail__data__element__head">Schedule minimum days</strong>
+                  <span className="test-detail__data__element__body">{`${test.scheduledDays} days`}</span>
+                </div>
+              </div>
+            </>
+          ))}
+        </>
+      ) : (
+        ''
+      )}
     </section>
   );
 }
 
 TestInfo.propTypes = {
   examName: PropTypes.string.isRequired,
-  option: PropTypes.shape.isRequired,
 };
 
 export default TestInfo;
